Separated example mpu6050 stm32 reg write / read functions into their own file
